removing coveralls to keep 1.8 support, moving helper to separate file
