+++ .travis.yml
@@ -1,3 +1,4 @@
 rvm:
+  - 1.8.7
   - 1.9.2
   - 1.9.3
