+++ Gemfile
@@ -1,5 +1,4 @@
 source "http://rubygems.org"
 
-gem 'coveralls', :require => false
 # Specify your gem's dependencies in har.gemspec
 gemspec
